requirements_distal_teaching.txt
numpy
matplotlib
tensorflow-gpu==1.12.0
pytransform3d

